+++ CHANGELOG.md
@@ -26,6 +26,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * The deprecated `--siblings` options for `jj split` has been removed.
   `jj split --parallel` can be used instead.
 
+* In colocated repos, the Git index now contains the changes from all parents
+  of the working copy instead of just the first parent (`HEAD`). 2-sided
+  conflicts from the merged parents are now added to the Git index as conflicts
+  as well.
+
 ### Deprecations
 
 ### New features